+++ packages/nestjs-endpoints/src/consts.ts
@@ -1,5 +1,8 @@
 export const endpointFileRegex = /\bendpoint\.(js|ts|mjs|cjs|mts)$/;
 
+export const routerModuleFileRegex =
+  /\brouter\.module\.(js|ts|mjs|cjs|mts)$/;
+
 export const settings: {
   endpoints: {
     file: string;