+++ packages/test/test-app-express-cjs/src/app.module.ts
@@ -15,6 +15,7 @@ import { Observable, catchError, throwError } from 'rxjs';
 import { ZodError } from 'zod';
 import { AuthModule } from './auth/auth.module';
 import { HelloService } from './endpoints/hello.service';
+import { RecipesManualEndpointsRouterModule } from './endpoints/recipes-manual/router.module';
 import {
   AppointmentRepositoryToken,
   IAppointmentRepository,
@@ -65,6 +66,7 @@ export class ZodErrorInterceptor implements NestInterceptor {
         HelloService,
       ],
     }),
+    RecipesManualEndpointsRouterModule,
     UserListModule,
   ],
   providers: [