+++ packages/test/test-app-express-cjs/src/endpoints/recipes/router.module.ts
@@ -1,0 +1,6 @@
+import { EndpointsRouterModule } from 'nestjs-endpoints';
+import { RecipesRepository } from './repository.service';
+
+export default EndpointsRouterModule.register({
+  providers: [RecipesRepository],
+});